QCamera2:Remove signalAPIResult for QCAMERA_SM_EVT_STOP_CAPTURE_CHANNEL
In NonZSL capture, we send QCAMERA_SM_EVT_STOP_CAPTURE_CHANNEL
from postproc thread as an API to Statemachine.
This is and async call so no one is waiting for signal after
completing this API

In failure case signal from this internal API call is blocking
previous call by overwriting m_apiResult value. Even with mutex
locks we have a race condition which could hit this.

Change-Id: I9ac663d4b703d38cd863d485951f3f1895a52120