QCamera2: Trigger rotation reprocess for livesnapshot also.
needReprocess was returning false for livesnapshots because
recording hint was set to true and feature_cpp mask was checked
in min required api instead of feature mask. So reprocess channel
was not getting added in some targets. Because of this livesnapshots
were not rotated properly. This change will remove the reprocess
requirement validation.

Change-Id: Iae2834b207ad73af648eb609a849c6bcc5fa48e4