Camera: Work around to support non-multimode architecture
To support JB and KK camera projects with common kernel, it
needs to restrict number of cameras to 2 as in L , max supported
camera declared as 4, in place of 2 earlier.
Introduced a setprop where user can set max number of cameras.
This'll resolve the issue for rear and front camera switch.

Change-Id: Ice40e5ed9d9ed59f6c44b92df666120efc9d424b