QCamera2: FOVControl: Fix a bug in case of using OEM fallback mechanism
Zoom direction is considered while moving between the WIDE, TRANSITION
and TELE states. The zoom can be constant when the fallback is initiated
by OEM library. As a result, the state transition from TELE to TRANSITION
doesn't occur. Ignore the zoom direction if the fallback is handled by
OEM library.

CRs-Fixed: 2039171

Change-Id: I0521544400b7023fcced9220a73091eca75ac131