QCamera2: FOV-control: Add configurable wait for the master switch
Add the wait for the master switch in terms of min number of frames.
With this configurable setting, the master cannot switch for the
configured number of frames after the camera starts streaming. So
when a camera comes out of LPM, it needs to wait for at least the
configured number of frames before assuming the master role.

CRs-Fixed: 2042222

Change-Id: I36d6ba0e175cac219975b66ae76dc08aa0f86a33