QCamera2: Ignore the focus distance for non-master when switching master
Currently FOVControl checks the AF distance and lux index of tele before
switching the master to tele. It takes 2-3 frames to update the AF
distance of non-master with master's AF info. This causes the latency in
master switch. As the non-master just copies the focus distance from the
current master, we can avoid this latency by ignoring the initial focus
distance reported by non-master. After 2-3 frames the focus distance for
non-master reflects the focus distance reported by the master.

CRs-Fixed: 2046050

Change-Id: I18668ea331fc6c5579683da0e3b4c6d0ed054dfd